+++ kernel/timeout.c
@@ -7,7 +7,7 @@
 #include <zephyr/kernel.h>
 #include <zephyr/spinlock.h>
 #include <ksched.h>
-#include <zephyr/timeout_q.h>
+#include <timeout_q.h>
 #include <zephyr/syscall_handler.h>
 #include <zephyr/drivers/timer/system_timer.h>
 #include <zephyr/sys_clock.h>
