+++ samples/kernel/metairq_dispatch/CMakeLists.txt
@@ -6,3 +6,7 @@ find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(metairq_dispatch)
 
 target_sources(app PRIVATE src/main.c src/msgdev.c)
+
+target_include_directories(app PRIVATE
+  ${ZEPHYR_BASE}/kernel/include
+  )
